validaciones en admin hasta cursos y refactorizado
